+++ pandashape/describers/__init__.py
@@ -1,3 +1,4 @@
 from .CorrelationDescriber import CorrelationDescriber
 from .Describer import Describer
+from .DistributionDescriber import DistributionDescriber
 from .GeneralDescriber import GeneralDescriber