+++ tests/smoke_describe.py
@@ -1,10 +1,10 @@
 import pandas as pd
 from pandashape import PandaShaper
-from pandashape.describers import GeneralDescriber, CorrelationDescriber
+from pandashape.describers import GeneralDescriber, DistributionDescriber, CorrelationDescriber
 
 df = pd.read_csv('./tests/assets/dataset.csv')
 df['AmountSpent_HighCorrelation'] = df['AmountSpent']
 df['Salary_HighCorr'] = df['Salary']
 
 shaper = PandaShaper(df)
-shaper.describe([GeneralDescriber, CorrelationDescriber])
+shaper.describe([GeneralDescriber, DistributionDescriber, CorrelationDescriber])