+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.0.5 
 - Corrected project homepage in setup.py
+- Added `DistributionDescriber` to describe skew and identify outliers.
 
 # 0.0.4
 - Updated README